requirements.txt
argon2-cffi==18.3.0
cffi==1.11.5
dj-database-url==0.5.0
Django==2.1.4
django-cors-headers==2.4.0
django-heroku==0.3.1
djangorestframework==3.9.0
djangorestframework-simplejwt==3.3
gunicorn==19.9.0
psycopg2-binary==2.7.6.1
pycparser==2.19
PyJWT==1.7.1
pytz==2018.7
six==1.12.0
whitenoise==4.1.2
